refactor: move module-scope test code into tests submodules
Separate the remaining inline test code that was not wrapped in a
`mod tests { ... }` block (so the earlier sweep did not catch it) into
sibling `tests.rs` submodules:

- fse, huff0 (mod + encoder): bare module-scope `#[test]` tail blocks.
- bit_reader_reverse: the `mod test` block plus the two `#[cfg(test)]`
  helper fns that only its tests used.
- sequence_section_decoder: module-scope tests plus the two nested
  `#[cfg(test)] mod` blocks; `super::` paths gain one level for the
  deeper nesting.
- hc/generator, hc/priceset: test fns left behind by the match-generator
  split. The `#[cfg(test)]` `start_matching` driver method stays with the
  struct (it reaches private fields); only the test cases move.

Two production helpers (`distribute_weights` / `redistribute_weights`,
called from `legacy_distributed_weights`) that sat among the encoder
tests are kept in production. Relative `include_bytes!` paths adjusted
for the deeper file locations. Suite unchanged at 824 cases; x86_64 +
BMI2 build cross-checked for the gated bit-reader helper.
